Added "5 kyu / Sort - one, three, two" first solution
While the main purpose of any code-education platform is to write code, I think the right approach to solve any catas is to write as much code as possible, avoiding the use of third-party libraries.
So it will be more useful to code your own functions instead of num2words and word2number.
I plan to update solution with my own libraries lat